requirements.txt
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
dill==0.3.6
Django==4.1.7
djangorestframework==3.14.0
dnspython==2.3.0
idna==3.4
imageio==2.26.0
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.1
lxml==4.9.2
networkx==3.0
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
Pillow==9.4.0
pymongo==3.12.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
requests==2.28.2
scikit-image==0.20.0
scikit-learn==1.2.1
six==1.16.0
soupsieve==2.4
sqlparse==0.2.4
threadpoolctl==3.1.0
tifffile==2023.2.28
tzdata==2022.7
urllib3==1.26.14
